+++ src/ilearnrw/languagetools/greek/GreekDictionary.java
@@ -81,7 +81,7 @@ public class GreekDictionary {
 					dictionary.put(row[0], new GreekWord(row[0], WordType.Unknown));
 				}
 				if (row.length == 4){
-					GreekWord w = new GreekWord(row[0], partOfSpeech(row[1]), row[2].trim(), Double.parseDouble(row[3]));
+					GreekWord w = new GreekWord(row[0], partOfSpeech(row[1]), row[2].trim(), Integer.parseInt(row[3]));
 
 					dictionary.put(row[0], w);
 				}
